requirements.txt
streamlit
langchain_groq 
langchain_community 
pypdf 
sentence_transformers
python-dotenv
chromadb
transformers
python-docx
docx2txt

